moved Life vendibility rules to sellable domain
